openvpn/requirements.txt
pyperclip
python-dotenv
requests
